refactor(Client): Moved specific functions to separate header file.
